updated
updated it because if we select no items or if we do not provide email in checkout page and click on place order then there will error or order failure but order id will be shown. if we track this id and email in tracker , it shows order placed which in reality was not. I guess this is a bug.